backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
pytest==8.2.0
httpx==0.27.0
langgraph==0.1.19
langchain-core==0.2.43
langchain-community==0.2.17
langchain-ollama==0.1.3

